+++ CHANGELOG.md
@@ -11,10 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added Uncategorized icon to map and legend
 - Added NFHL layer and legend for NFHL sublayer
+- Added NWS Radar layer
 
 ### Changed
 
--
+- RP icons visible in legend only if on the map
 
 ### Fixed